// TempleTreasure/out/lib/game_library.d.ts
export declare enum Color {
	AliceBlue = "#F0F8FF",
	AntiqueWhite = "#FAEBD7",
	Aqua = "#00FFFF",
	Aquamarine = "#7FFFD4",
	Azure = "#F0FFFF",
	Beige = "#F5F5DC",
	Bisque = "#FFE4C4",
	Black = "#000000",
	BlanchedAlmond = "#FFEBCD",
	Blue = "#0000FF",
	BlueViolet = "#8A2BE2",
	Brown = "#A52A2A",
	BurlyWood = "#DEB887",
	CadetBlue = "#5F9EA0",
	Chartreuse = "#7FFF00",
	Chocolate = "#D2691E",
	Coral = "#FF7F50",
	CornflowerBlue = "#6495ED",
	Cornsilk = "#FFF8DC",
	Crimson = "#DC143C",
	Cyan = "#00FFFE",
	DarkBlue = "#00008B",
	DarkCyan = "#008B8B",
	DarkGoldenrod = "#B8860B",
	DarkGray = "#A9A9A9",
	DarkGreen = "#006400",
	DarkKhaki = "#BDB76B",
	DarkMagenta = "#8B008B",
	DarkOliveGreen = "#556B2F",
	DarkOrange = "#FF8C00",
	DarkOrchid = "#9932CC",
	DarkRed = "#8B0000",
	DarkSalmon = "#E9967A",
	DarkSeaGreen = "#8FBC8F",
	DarkSlateBlue = "#483D8B",
	DarkSlateGray = "#2F4F4F",
	DarkTurquoise = "#00CED1",
	DarkViolet = "#9400D3",
	DeepPink = "#FF1493",
	DeepSkyBlue = "#00BFFF",
	DimGray = "#696969",
	DodgerBlue = "#1E90FF",
	Firebrick = "#B22222",
	FloralWhite = "#FFFAF0",
	ForestGreen = "#228B22",
	Fuchsia = "#FF00FE",
	Gainsboro = "#DCDCDC",
	GhostWhite = "#F8F8FF",
	Gold = "#FFD700",
	Goldenrod = "#DAA520",
	Gray = "#808080",
	Green = "#008000",
	GreenYellow = "#ADFF2F",
	Honeydew = "#F0FFF0",
	HotPink = "#FF69B4",
	IndianRed = "#CD5C5C",
	Indigo = "#4B0082",
	Ivory = "#FFFFF0",
	Khaki = "#F0E68C",
	Lavender = "#E6E6FA",
	LavenderBlush = "#FFF0F5",
	LawnGreen = "#7CFC00",
	LemonChiffon = "#FFFACD",
	LightBlue = "#ADD8E6",
	LightCoral = "#F08080",
	LightCyan = "#E0FFFF",
	LightGoldenrodYellow = "#FAFAD2",
	LightGreen = "#90EE90",
	LightGray = "#D3D3D3",
	LightPink = "#FFB6C1",
	LightSalmon = "#FFA07A",
	LightSeaGreen = "#20B2AA",
	LightSkyBlue = "#87CEFA",
	LightSlateGray = "#778899",
	LightSteelBlue = "#B0C4DE",
	LightYellow = "#FFFFE0",
	Lime = "#00FF00",
	LimeGreen = "#32CD32",
	Linen = "#FAF0E6",
	Magenta = "#FF00FF",
	Maroon = "#800000",
	MediumAquamarine = "#66CDAA",
	MediumBlue = "#0000CD",
	MediumOrchid = "#BA55D3",
	MediumPurple = "#9370DB",
	MediumSeaGreen = "#3CB371",
	MediumSlateBlue = "#7B68EE",
	MediumSpringGreen = "#00FA9A",
	MediumTurquoise = "#48D1CC",
	MediumVioletRed = "#C71585",
	MidnightBlue = "#191970",
	MintCream = "#F5FFFA",
	MistyRose = "#FFE4E1",
	Moccasin = "#FFE4B5",
	NavajoWhite = "#FFDEAD",
	Navy = "#000080",
	OldLace = "#FDF5E6",
	Olive = "#808000",
	OliveDrab = "#6B8E23",
	Orange = "#FFA500",
	OrangeRed = "#FF4500",
	Orchid = "#DA70D6",
	PaleGoldenrod = "#EEE8AA",
	PaleGreen = "#98FB98",
	PaleTurquoise = "#AFEEEE",
	PaleVioletRed = "#DB7093",
	PapayaWhip = "#FFEFD5",
	PeachPuff = "#FFDAB9",
	Peru = "#CD853F",
	Pink = "#FFC0CB",
	Plum = "#DDA0DD",
	PowderBlue = "#B0E0E6",
	Purple = "#800080",
	Red = "#FF0000",
	RosyBrown = "#BC8F8F",
	RoyalBlue = "#4169E1",
	SaddleBrown = "#8B4513",
	Salmon = "#FA8072",
	SandyBrown = "#F4A460",
	SeaGreen = "#2E8B57",
	SeaShell = "#FFF5EE",
	Sienna = "#A0522D",
	Silver = "#C0C0C0",
	SkyBlue = "#87CEEB",
	SlateBlue = "#6A5ACD",
	SlateGray = "#708090",
	Snow = "#FFFAFA",
	SpringGreen = "#00FF7F",
	SteelBlue = "#4682B4",
	Tan = "#D2B48C",
	Teal = "#008080",
	Thistle = "#D8BFD8",
	Tomato = "#FF6347",
	Turquoise = "#40E0D0",
	Violet = "#EE82EE",
	Wheat = "#F5DEB3",
	White = "#FFFFFF",
	WhiteSmoke = "#F5F5F5",
	Yellow = "#FFF000",
	YellowGreen = "#9ACD32"
}
export declare class ContentManager {
	private loadedAssets;
	private disposableAssets;
	private static contentExtension;
	private serviceProviders;
	private rootDirectory;
	private isRootDirectoryAbsolute;
	fullRootDirectory: string;
	private byteBuffer;
	private weakByteBuffers;
	private readAssetRecurseCount;
	constructor();
	get RootDirectory(): string;
	set RootDirectory(value: string);
}
export declare class Controls {
	m_left: boolean;
	m_right: boolean;
	m_up: boolean;
	m_down: boolean;
	m_fire: boolean;
	m_spacePressed: boolean;
	m_firePressed: boolean;
	m_divePressed: boolean;
	m_reversePressed: boolean;
	m_return: boolean;
	m_enter: boolean;
	m_lcontrol: boolean;
	m_record: boolean;
	m_play: boolean;
	constructor();
	get left(): boolean;
	set left(value: boolean);
	get right(): boolean;
	set right(value: boolean);
	get up(): boolean;
	set up(value: boolean);
	get down(): boolean;
	set down(value: boolean);
	get lcontrolPressed(): boolean;
	set lcontrolPressed(value: boolean);
	get enterPressed(): boolean;
	set enterPressed(value: boolean);
	get returnPressed(): boolean;
	set returnPressed(value: boolean);
	get spacePressed(): boolean;
	set spacePressed(value: boolean);
}
export declare class TimeSpan {
	private TicksPerMillisecond;
	private MillisecondsPerTick;
	private ticksPerSecond;
	private SecondsPerTick;
	private TicksPerMinute;
	private MinutesPerTick;
	private TicksPerHour;
	private HoursPerTick;
	private TicksPerDay;
	private DaysPerTick;
	private MillisPerSecond;
	private MillisPerMinute;
	private MillisPerHour;
	private MillisPerDay;
	private MaxValue_Int64;
	private MinValue_Int64;
	private MaxSeconds;
	private MinSeconds;
	private MaxMilliSeconds;
	private MinMilliSeconds;
	private TicksPerTenthSecond;
	private _ticks;
	constructor(ticks: number);
	get Zero(): TimeSpan;
	get Ticks(): number;
	set Ticks(value: number);
	get TicksPerSecond(): number;
	get Days(): number;
	get Hours(): number;
	get Milliseconds(): number;
	get Minutes(): number;
	get Seconds(): number;
	get TotalDays(): number;
	get TotalHours(): number;
	get TotalMilliseconds(): number;
	get TotalMinutes(): number;
	get TotalSeconds(): number;
	Add(ts: TimeSpan): TimeSpan;
	Subtract(ts: TimeSpan): TimeSpan;
	LessThan(ts: number): boolean;
	MoreThan(ts: number): boolean;
	EqualTo(ts: number): boolean;
	Compare(t1: TimeSpan, t2: TimeSpan): number;
	CompareTo(value: Object): number;
	FromMinutes(value: number): TimeSpan;
	FromSeconds(value: number): TimeSpan;
	private Interval;
	SetTicks(value: number): void;
	FromTicks(value: number): TimeSpan;
	FromMilliseconds(value: number): TimeSpan;
}
export declare class GameTime {
	totalGameTime: TimeSpan;
	elapsedGameTime: TimeSpan;
	isRunningSlowly: boolean;
	constructor();
	set TotalGameTime(value: TimeSpan);
	get TotalGameTime(): TimeSpan;
	set ElapsedGameTime(value: TimeSpan);
	get ElapsedGameTime(): TimeSpan;
	set IsRunningSlowly(value: boolean);
	get IsRunningSlowly(): boolean;
}
export declare class Point {
	static Zero: any;
	private m_x;
	private m_y;
	constructor(x: number, y: number);
	get X(): number;
	set X(value: number);
	get Y(): number;
	set Y(value: number);
}
export declare class Rectangle {
	X: number;
	Y: number;
	Width: number;
	Height: number;
	static get Empty(): Rectangle;
	get Left(): number;
	get Right(): number;
	get Top(): number;
	get Bottom(): number;
	constructor(x: number, y: number, width: number, height: number);
	get Center(): Point;
	Contains(value: Point): boolean;
	Intersects(value: Rectangle): boolean;
}
export declare enum SpriteEffects {
	None = 0
}
export type Texture2D = HTMLImageElement;
export declare class SpriteBatch {
	private canvas;
	private ctx;
	get Canvas(): HTMLCanvasElement;
	constructor(non?: any);
	Draw(texture: Texture2D, param2: any, param3?: number | Color | Rectangle, param4?: any, param5?: any, param6?: any, param7?: any, param8?: any, param9?: any): void;
	DrawString(font: any, value: any, position: any, color: any): void;
	DrawButton(button: any): void;
	Begin(): void;
	Clear(col: string): void;
	End(): void;
}
export declare class Game {
	private maximumElapsedTime;
	private isFixedTimeStep;
	private int_MaxValue;
	private updatesSinceRunningSlowly1;
	private updatesSinceRunningSlowly2;
	private isActive;
	private exitRequested;
	private inactiveSleepTime;
	private isMouseVisible;
	private inRun;
	private endRunRequired;
	private clock;
	private totalGameTime;
	private targetElapsedTime;
	private accumulatedElapsedGameTime;
	private lastFrameElapsedGameTime;
	private drawRunningSlowly;
	private doneFirstUpdate;
	private doneFirstDraw;
	private forceElapsedTimeToZero;
	private suppressDraw;
	private content;
	protected gameTime: GameTime;
	protected spriteBatch: SpriteBatch;
	constructor();
	Run(gameTime: GameTime): void;
	get GraphicsDevice(): SpriteBatch;
	get Content(): ContentManager;
	set Content(value: ContentManager);
	get MaxElapsedTime(): TimeSpan;
	set MaxElapsedTime(value: TimeSpan);
	get IsActive(): boolean;
	get IsMouseVisible(): boolean;
	set IsMouseVisible(value: boolean);
	get TargetElapsedTime(): TimeSpan;
	set TargetElapsedTime(value: TimeSpan);
	get IsFixedTimeStep(): boolean;
	set IsFixedTimeStep(value: boolean);
	RunGame(useBlockingRun: boolean): void;
	private get ShouldExit();
	Tick(): void;
	private DrawFrame;
	protected Draw(gameTime: GameTime): void;
	protected Update(gameTime: GameTime): void;
	protected Initialize(caller: any): void;
}
// ...
export declare class Vector2 {
	static Zero: any;
	private m_x;
	private m_y;
	constructor(x: number, y: number);
	get X(): number;
	set X(value: number);
	get Y(): number;
	set Y(value: number);
	get Zero(): Vector2;
	LengthSquared(): number;
}
export declare class RectangleExtensions {
	static GetIntersectionDepth(rectA: any, rectB: any): Vector2;
	static GetBottomCenter(rect: any): Vector2;
}
export declare class TexturedButton {
	x: number;
	y: number;
	width: number;
	height: number;
	private image;
	private onClick;
	constructor(x: number, y: number, width: number, height: number, imageSrc: string, onClick: () => void);
	draw(ctx: CanvasRenderingContext2D): void;
	containsPoint(mouseX: number, mouseY: number): boolean;
	click(mouseX: number, mouseY: number): void;
}